cloud_function/requirements.txt
numpy==1.24.3
pandas==2.0.3
google-cloud-storage==2.9.0
google-cloud-bigquery==3.11.4
google-cloud-logging==3.5.0
pytz==2023.3
requests==2.31.0
retry==0.9.2
pyarrow==14.0.2 